Creating abstract class character and Updating the ones that already existed
