+++ src/Library/Characters/Archer.cs
@@ -1,21 +1,16 @@
 namespace RoleplayGame
 {
-    public class Archer
+    public class Archer : Character
     {
-        private int health = 100;
-
-        public Archer(string name)
+        public Archer(string name) : base(name)
         {
-            this.Name = name;
         }
-
-        public string Name { get; set; }
         
         public Bow Bow { get; set; }
 
         public Helmet Helmet { get; set; }
 
-        public int AttackValue
+        public override int AttackValue
         {
             get
             {
@@ -23,7 +18,7 @@ namespace RoleplayGame
             }
         }
 
-        public int DefenseValue
+        public override int DefenseValue
         {
             get
             {
@@ -31,29 +26,5 @@ namespace RoleplayGame
             }
         }
 
-        public int Health
-        {
-            get
-            {
-                return this.health;
-            }
-            private set
-            {
-                this.health = value < 0 ? 0 : value;
-            }
-        }
-
-        public void ReceiveAttack(int power)
-        {
-            if (this.DefenseValue < power)
-            {
-                this.Health -= power - this.DefenseValue;
-            }
-        }
-
-        public void Cure()
-        {
-            this.Health = 100;
-        }
     }
 }
