+++ src/Library/Characters/Dwarf.cs
@@ -1,15 +1,10 @@
 namespace RoleplayGame
 {
-    public class Dwarf
+    public class Dwarf : Character
     {
-        private int health = 100;
-
-        public Dwarf(string name)
+        public Dwarf(string name) : base(name)
         {
-            this.Name = name;
         }
-
-        public string Name { get; set; }
 
         public Axe Axe { get; set; }
 
@@ -17,7 +12,7 @@ namespace RoleplayGame
 
         public Helmet Helmet { get; set; }
 
-        public int AttackValue
+        public override int AttackValue
         {
             get
             {
@@ -25,37 +20,12 @@ namespace RoleplayGame
             }
         }
 
-        public int DefenseValue
+        public override int DefenseValue
         {
             get
             {
                 return Shield.DefenseValue + Helmet.DefenseValue;
             }
-        }
-
-        public int Health
-        {
-            get
-            {
-                return this.health;
-            }
-            private set
-            {
-                this.health = value < 0 ? 0 : value;
-            }
-        }
-
-        public void ReceiveAttack(int power)
-        {
-            if (this.DefenseValue < power)
-            {
-                this.Health -= power - this.DefenseValue;
-            }
-        }
-
-        public void Cure()
-        {
-            this.Health = 100;
         }
     }
 }
