+++ src/Library/Characters/Wizard.cs
@@ -1,21 +1,16 @@
 namespace RoleplayGame
 {
-    public class Wizard
+    public class Wizard : Character
     {
-        private int health = 100;
-
-        public Wizard(string name)
+        public Wizard(string name) : base(name)
         {
-            this.Name = name;
         }
-
-        public string Name { get; set; }
 
         public SpellsBook SpellsBook { get; set; }
 
         public Staff Staff { get; set; }
 
-        public int AttackValue
+        public override int AttackValue
         {
             get
             {
@@ -23,37 +18,12 @@ namespace RoleplayGame
             }
         }
 
-        public int DefenseValue
+        public override int DefenseValue
         {
             get
             {
                 return SpellsBook.DefenseValue + Staff.DefenseValue;
             }
-        }
-
-        public int Health
-        {
-            get
-            {
-                return this.health;
-            }
-            private set
-            {
-                this.health = value < 0 ? 0 : value;
-            }
-        }
-
-        public void ReceiveAttack(int power)
-        {
-            if (this.DefenseValue < power)
-            {
-                this.Health -= power - this.DefenseValue;
-            }
-        }
-
-        public void Cure()
-        {
-            this.Health = 100;
         }
     }
 }
